Implemented update so that when cards are moved and the page is refresh, it is still saved
