+++ app/views/selections/index.html.erb
@@ -43,7 +43,7 @@
 
           <% selection = Selection.where(user: current_user, time_of_day: day[1], date: day[0]).first %>
           <% if selection %>
-            <div data-selection-id="<%=@counter%>" class="card-cell card-itinerary" style="background-image: url('<%=selection.activity.image_url%>');">
+            <div data-selection-id="<%=selection.id%>" class="card-cell card-itinerary" style="background-image: url('<%=selection.activity.image_url%>');">
 
               <span class="card-none"> <i class="fas fa-random"></i></span>
               <p class><%= selection.activity.name %></p>
